+++ .eslintrc.cjs
@@ -66,6 +66,9 @@ module.exports = {
         // TODO: part of "stylistic" rules, remove explicit activation when that lands
         '@typescript-eslint/no-empty-function': 'error',
         '@typescript-eslint/no-empty-interface': 'error',
+
+        // not needed to be enforced for TS
+        'import/namespace': 'off',
       },
     },
     {
@@ -160,6 +163,7 @@ module.exports = {
         'e2e/failures/macros.js',
         'e2e/test-in-root/*.js',
         'e2e/test-match/test-suites/*',
+        'e2e/test-match-default/dot-spec-tests/*',
         'packages/test-utils/src/ConditionalTest.ts',
       ],
       env: {'jest/globals': true},
@@ -599,6 +603,7 @@ module.exports = {
     'unicorn/explicit-length-check': 'error',
     'unicorn/no-array-for-each': 'error',
     'unicorn/no-negated-condition': 'error',
+    'unicorn/numeric-separators-style': 'error',
     'unicorn/prefer-default-parameters': 'error',
     'unicorn/prefer-includes': 'error',
     'unicorn/template-indent': 'error',