+++ e2e/detect-open-handles/__tests__/in-done-function.js
@@ -6,7 +6,7 @@
  */
 
 test('something', done => {
-  setTimeout(() => {}, 10000);
+  setTimeout(() => {}, 10_000);
   expect(true).toBe(true);
   done();
 });