+++ packages/jest-each/src/__tests__/array.test.ts
@@ -283,12 +283,12 @@ describe('jest-each', () => {
         const eachObject = each.withGlobal(globalTestMocks)([['hello']]);
 
         const testFunction = get(eachObject, keyPath);
-        testFunction('some test', noop, 10000);
+        testFunction('some test', noop, 10_000);
         const globalMock = get(globalTestMocks, keyPath);
         expect(globalMock).toHaveBeenCalledWith(
           'some test',
           expect.any(Function),
-          10000,
+          10_000,
         );
       });